refactor ref-finance
